requirements.txt
flake8
mock
nose
six
requests>=2.0.0
PyJWT>=1.4.2
twine
PyNaCl
